Test: api에 의존적인 테스트 mocking
